+++ test/e2e/suites/allocate/node_order/node_order_suite_test.go
@@ -15,6 +15,7 @@ import (
 
 var _ = DescribeAffinitySpecs()
 var _ = DescribeResourceTypeSpecs()
+var _ = DescribeFillNodeSpecs()
 
 func TestNodeOrder(t *testing.T) {
 	utils.SetLogger()